+++ components/layout/Footer.js
@@ -1,11 +1,9 @@
 import React from 'react';
-
 import { styled } from '@mui/material/styles';
 import { Divider, Typography } from '@mui/material/';
-
 import packageJson from '../../package.json';
 import SocialMediaLink from '../socialMedia/SocialMediaLink';
-import { isMobileOnly } from 'react-device-detect';
+import { MOBILE, MOBILE_SM } from '../../utils/constants';
 
 const PREFIX = 'Footer';
 
@@ -13,10 +11,11 @@ const classes = {
   footerDiv: `${PREFIX}-footerDiv`,
   container: `${PREFIX}-container`,
   divider: `${PREFIX}-divider`,
-  whiteText: `${PREFIX}-whiteText`
+  whiteText: `${PREFIX}-whiteText`,
+  author: `${PREFIX}-author`
 };
 
-const Root = styled('div')((
+const Root = styled('footer')((
   {
     theme
   }
@@ -26,9 +25,10 @@ const Root = styled('div')((
     padding: "4em 0 2em 0",
     marginTop: "2em",
     textAlign: "center",
-    letterSpacing: isMobileOnly ? 0 : 3,
-    '@media (max-width:900px)': {
+    letterSpacing: 3,
+    [`@media (max-width:${MOBILE})`]: {
       padding: "2em 0",
+      letterSpacing: 0
     }
   },
 
@@ -37,7 +37,7 @@ const Root = styled('div')((
     display: "flex",
     justifyContent: "space-between",
     padding: "0 4em",
-    '@media (max-width:900px)': {
+    [`@media (max-width:${MOBILE})`]: {
       flexWrap: "wrap",
       padding: "0 2em",
     }
@@ -47,13 +47,20 @@ const Root = styled('div')((
     margin: "2em 4em",
     backgroundColor: theme.palette.primary.contrastText,
     opacity: 0.5,
-    '@media (max-width:900px)': {
+    [`@media (max-width:${MOBILE})`]: {
       margin: "2em 1em 0 1em",
     }
   },
 
   [`& .${classes.whiteText}`]: {
     color: theme.palette.primary.contrastText,
+  },
+
+  [`& .${classes.author}`]: {
+    letterSpacing: 3,
+    [`@media (max-width:${MOBILE_SM})`]: {
+      letterSpacing: 0
+    }
   }
 }));
 
@@ -80,7 +87,7 @@ export default function Footer() {
           </Typography>
         </div>
         <div>
-          <Typography variant="caption" className={classes.whiteText} style={{ letterSpacing: isMobileOnly ? 0 : 3 }}>
+          <Typography variant="caption" className={`${classes.whiteText} ${classes.author}`}>
             &copy; {currentYear} Anika Mlodzianowski
           </Typography>
         </div>