+++ components/layout/Main.js
@@ -1,11 +1,11 @@
 import React from "react";
 import { styled } from '@mui/material/styles';
-import { isMobile } from 'react-device-detect';
 import Image from 'next/image'
 import { Container } from '@mui/material/';
 import Illustration from "./Illustration";
 import Trails from './Trails';
 import Scroll from "./Scroll";
+import { MOBILE, MOBILE_SM } from "../../utils/constants";
 
 const PREFIX = 'Main';
 
@@ -24,28 +24,27 @@ const Root = styled('div')(() => ({
     margin: "2em",
     padding: "4em 2em",
     height: 'calc(100vh - 64px)',
-    '@media (max-width:900px)': {
+    [`@media (max-width:${MOBILE})`]: {
       flexWrap: "wrap",
       height: '100%',
       margin: "1em",
       padding: "2em",
     },
-  },
-
-  [`& .${classes.containerMobile}`]: {
-    height: 'auto',
-    flexWrap: "wrap",
-    margin: "1em",
-    padding: "0em",
-    flexDirection: 'column',
-    justifyContent: 'space-between',
+    [`@media (max-width:${MOBILE_SM})`]: {
+      height: 'auto',
+      flexWrap: "wrap",
+      margin: "1em",
+      padding: "0em",
+      flexDirection: 'column',
+      justifyContent: 'space-between',
+    }
   },
 
   [`& .${classes.trailsContainer}`]: {
     margin: "2em auto",
     paddingLeft: "4em",
     width: '50vw',
-    '@media (max-width:900px)': {
+    [`@media (max-width:${MOBILE})`]: {
       height: "unset",
       padding: 0,
       margin: "2em",
@@ -56,21 +55,22 @@ const Root = styled('div')(() => ({
   [`& .${classes.illustration}`]: {
     overflow: "visible",
     textAlign: "center",
-    // height: "auto",
-    // maxWidth: '50vw',
-    // minWidth: 300,
     width: '100%',
     margin: '0 auto',
-    '@media (max-width:900px)': {
+    [`@media (max-width:${MOBILE})`]: {
       maxWidth: '80vw',
     },
   },
 
   [`& .${classes.icon}`]: {
+    display: 'none',
     position: 'relative',
-    maxWidth: "-webkit-fill-available",
-    height: "20em",
-    maxHeight: '40vh',
+    [`@media (max-width:${MOBILE_SM})`]: {
+      display: 'block',
+      maxWidth: "-webkit-fill-available",
+      height: "20em",
+      maxHeight: '40vh',
+    }
   }
 }));
 
@@ -84,16 +84,17 @@ export default function Main() {
 
   return (
     <Root>
-      <div className={isMobile ? classes.containerMobile : classes.container}>
+      <div className={classes.container}>
         <div className={classes.text}>
           <Container maxWidth="lg" className={classes.trailsContainer}>
             <Trails />
           </Container>
         </div>
         <div className={classes.illustration}>
-          {isMobile ? illustration : <Illustration />}
+          {illustration}
+          <Illustration />
         </div>
-        {isMobile ? null : <Scroll />}
+        <Scroll />
       </div>
     </Root>
   );