+++ utils/constants.js
@@ -1,0 +1,2 @@
+export const MOBILE_SM = '400px';
+export const MOBILE = '1000px';